refactor: `OAuthRestClient` 인터페이스를 클래스로 변경
